# Improve ISAAC performance (take 2)