This is a new try instead of https://github.com/dhardy/rand/pull/36. Now fills the output buffer in reverse, so `fill_bytes` keeps working exactly as before.

It was now possible to move code shared between `IsaacRng` and `ChaChaRng` to a separate `rand_core::impl::fill_via_u*_chunks` function. All the unsafe code is contained therein.

The trick with 32-bit indexing to make `isaac64::next_u32` faster worked best with reading the results backwards. Now I had to think of something else, and it isn't pretty...