@pitdicker 2017-11-11
Sorry for all the pushes.

I temporarily merged https://github.com/dhardy/rand/pull/44, and big-endian tests [are green](https://travis-ci.org/dhardy/rand/builds/300522345?utm_source=github_status&utm_medium=notification) (not on first try though).