@pitdicker 2017-11-11
It seems our comments have crossed. I also hand-calculated the results. Is the one in the commit ok for you?